feat: Refactor panda to use source + header file structure for driver modules
